+++ src/main/go/domain/infra/repository/allocation_plan_repository_model.go
@@ -8,11 +8,11 @@ import (
 )
 
 type plannedAllocationJoinedRowDTS struct {
-	AllocationPlanId     int
+	AllocationPlanId     int64
 	Name                 string
 	Type                 allocation.PlanType
 	PlannedExecutionDate sqlext.NullTime
-	PlannedAllocationId  int
+	PlannedAllocationId  int64
 	HierarchicalId       sqlext.NullStringSlice
 	CashReserve          bool
 	SliceSizePercentage  decimal.Decimal
@@ -20,7 +20,7 @@ type plannedAllocationJoinedRowDTS struct {
 
 func mapPlannedAllocationRows(rows []plannedAllocationJoinedRowDTS) ([]*domain.AllocationPlan, error) {
 
-	var allocationPlanCacheMap = make(map[int]*domain.AllocationPlan)
+	var allocationPlanCacheMap = make(map[int64]*domain.AllocationPlan)
 	var allocationPlans = make([]*domain.AllocationPlan, 0)
 
 	for _, row := range rows {
@@ -36,7 +36,7 @@ func mapPlannedAllocationRows(rows []plannedAllocationJoinedRowDTS) ([]*domain.A
 
 func mapPlannedAllocationRow(
 	row *plannedAllocationJoinedRowDTS,
-	allocationPlanCacheMap map[int]*domain.AllocationPlan,
+	allocationPlanCacheMap map[int64]*domain.AllocationPlan,
 ) *domain.AllocationPlan {
 
 	allocationPlanUnit := buildPlannedAllocationFromRow(row)
@@ -54,7 +54,7 @@ func mapPlannedAllocationRow(
 
 func buildPlannedAllocationFromRow(rowDTS *plannedAllocationJoinedRowDTS) *domain.PlannedAllocation {
 	return &domain.PlannedAllocation{
-		Id:                  rowDTS.PlannedAllocationId,
+		Id:                  int64(rowDTS.PlannedAllocationId),
 		HierarchicalId:      rowDTS.HierarchicalId.ToStringSlice(),
 		CashReserve:         rowDTS.CashReserve,
 		SliceSizePercentage: rowDTS.SliceSizePercentage,
@@ -68,7 +68,7 @@ func buildAllocationPlanFromRow(
 
 	var allocationPlan = domain.AllocationPlan{
 		AllocationPlanIdentifier: domain.AllocationPlanIdentifier{
-			Id:   rowDTS.AllocationPlanId,
+			Id:   int64(rowDTS.AllocationPlanId),
 			Name: rowDTS.Name,
 		},
 		PlanType:             rowDTS.Type,
